products

+++ nu3Task/Entities/Image.cs
@@ -8,8 +8,8 @@ namespace nu3Task.Entities
     public partial class Image
     {
         public int Id { get; set; }
-        public int ImageId { get; set; }
-        public int ProductId { get; set; }
+        public long ImageId { get; set; }
+        public long ProductId { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime UpdatedAt { get; set; }
         public int Width { get; set; }
